sample submission

+++ pegr/grails-app/domain/pegr/Sample.groovy
@@ -6,10 +6,10 @@ class Sample {
 	Antibody antibody
     String antibodyNotes
 	Target target
-	Long requestedTagNumber
-	Float chromosomeAmount // in ug
-	Long cellNumber // in M
-	Float volume //ul per aliqu used for the assay
+	Double requestedTagNumber
+	Double chromosomeAmount // in ug
+	Double cellNumber // in M
+	Double volume //ul per aliqu used for the assay
 	String publicationReference
     SampleStatus status
     Date date
